fixed an issue where default tooltip background could not be set from a Style
