+++ UwpView/DefaultLegend.xaml.cs
@@ -30,20 +30,6 @@ using LiveCharts.Uwp.Components.MultiBinding;
 namespace LiveCharts.Uwp
 {
     /// <summary>
-    /// 
-    /// </summary>
-    public interface IChartLegend
-    {
-        /// <summary>
-        /// Gets or sets the series.
-        /// </summary>
-        /// <value>
-        /// The series.
-        /// </value>
-        List<SeriesViewModel> Series { get; set; }
-    }
-
-    /// <summary>
     /// The default legend control, by default a new instance of this control is created for every chart that requires a legend.
     /// </summary>
     public partial class DefaultLegend : IChartLegend
