+++ WpfView/DefaultTooltip.xaml
@@ -33,9 +33,6 @@
              mc:Ignorable="d" 
              d:DataContext="{d:DesignInstance wpf:DefaultTooltip}"
              BorderThickness="1.3">
-    <UserControl.Background>
-        <SolidColorBrush Color="#202020" Opacity=".8" />
-    </UserControl.Background>
     <UserControl.Resources>
         <Style TargetType="TextBlock">
             <Setter Property="Foreground" Value="{Binding Foreground}"></Setter>
